# docs/swagger.yaml
basePath: /
definitions:
...
  model.ErrorWrapper:
    properties:
      access_token:
        type: string
      code:
        type: integer
      data: {}
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/model.PaginationPage'
    type: object
...
  model.LoginRequest:
    properties:
      password:
        maxLength: 255
        minLength: 3
        type: string
      username:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  model.PaginationPage:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  model.RegisterRequest:
    properties:
      name:
        maxLength: 255
        type: string
      password:
        maxLength: 255
        minLength: 3
        type: string
      username:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
...
  model.UserResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      username:
        type: string
    type: object
  model.UserResponseWrapper:
    properties:
      access_token:
        type: string
      code:
        type: integer
      data:
        $ref: '#/definitions/model.UserResponse'
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/model.PaginationPage'
    type: object
host: 192.168.110.65:3000
info:
  contact: {}
  description: Enter your API Key as X-API-Key header value
  title: Aino Medical API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Login a user with the provided credentials.
      parameters:
      - description: Login User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
      summary: Login a user
      tags:
      - Users
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user with the provided details.
      parameters:
      - description: Create User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.RegisterRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorWrapper'
      summary: Create a new user
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"